ts-migration: convert healchecks, progresses, pullrequests, staging and trading in typescript
