+++ routes/index.ts
@@ -7,14 +7,14 @@ app.use("/auth", require("./auth.ts"));
 app.use("/badges", require("./badges.js"));
 app.use("/challenges", require("./challenges.ts"));
 app.use("/contributions", require("./contributions.ts"));
-app.use("/healthcheck", require("./healthCheck.js"));
+app.use("/healthcheck", require("./healthCheck.ts"));
 app.use("/logs", require("./logs.js"));
 app.use("/members", require("./members.js"));
-app.use("/pullrequests", require("./pullrequests.js"));
+app.use("/pullrequests", require("./pullrequests.ts"));
 app.use("/stocks", require("./stocks.js"));
 app.use("/tasks", require("./tasks.js"));
 app.use("/taskRequests", require("./taskRequests"));
-app.use("/trade", require("./trading.js"));
+app.use("/trade", require("./trading.ts"));
 app.use("/users/status", require("./userStatus.js"));
 app.use("/users", require("./users.js"));
 app.use("/profileDiffs", require("./profileDiffs.js"));
@@ -28,8 +28,8 @@ app.use("/external-accounts", require("./external-accounts.js"));
 app.use("/events", require("./events.js"));
 app.use("/discord-actions", require("./discordactions.js"));
 app.use("/issues", require("./issues.ts"));
-app.use("/progresses", require("./progresses.js"));
+app.use("/progresses", require("./progresses.ts"));
 app.use("/monitor", require("./monitor.js"));
-app.use("/staging", require("./staging.js"));
+app.use("/staging", require("./staging.ts"));
 app.use("/goals", require("./goals.ts"));
 module.exports = app;
