+++ routes/pullrequests.ts
@@ -1,6 +1,6 @@
-const express = require("express");
+import express from "express";
+import pullRequest from "../controllers/pullRequests";
 const router = express.Router();
-const pullRequest = require("../controllers/pullRequests");
 
 router.get("/open", pullRequest.getOpenPRs);
 router.get("/stale", pullRequest.getStalePRs);
