--- routes/trading.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const authenticate = require("../middlewares/authenticate");
-const { trade } = require("../controllers/trading");
-const { newTrade } = require("../middlewares/validators/trading");
-
-router.post("/stock/new/self", authenticate, newTrade, trade);
-
-module.exports = router;
